Improve areas of the bot and update deps & faq

+++ src/events/messageUpdate.ts
@@ -4,14 +4,24 @@ export default {
     name: 'messageUpdate',
     execute: async(client:TClient, oldMsg:Discord.Message, newMsg:Discord.Message)=>{
         if (!client.config.botSwitches.logs) return;
-        const disabledChannels = ['548032776830582794', '541677709487505408']
-        if (oldMsg.guild?.id != client.config.mainServer.id) return;
-        if (oldMsg.author == null) return;
-        if (oldMsg?.author.bot) return;
-        if (oldMsg.partial) return;
-        if (newMsg.partial) return;
-        if (!newMsg.member) return;
-        if (disabledChannels.includes(newMsg.channelId)) return;
+        const disabledChannels = [
+            '548032776830582794', '541677709487505408'
+        ]
+        if (
+            oldMsg.guild?.id != client.config.mainServer.id
+            || oldMsg.author == null
+            || oldMsg?.author.bot
+            || oldMsg.partial
+            || newMsg.partial
+            || !newMsg.member
+            || disabledChannels.includes(newMsg.channelId)
+        ) return;
+        // if (oldMsg.author == null) return;
+        // if (oldMsg?.author.bot) return;
+        // if (oldMsg.partial) return;
+        // if (newMsg.partial) return;
+        // if (!newMsg.member) return;
+        // if (disabledChannels.includes(newMsg.channelId)) return;
         const msgarr = newMsg.content.toLowerCase().split(' ');
         if (client.bannedWords._content.some((word:string)=>msgarr.includes(word)) && (!client.isStaff(newMsg.member))) newMsg.delete();
         if (newMsg.content === oldMsg.content) return;
